Used map function to render ratings
